+++ examples/nextjs/server.js
@@ -5,22 +5,27 @@ const port = parseInt(process.env.PORT, 10) || 3000;
 const dev = process.env.NODE_ENV !== 'production';
 const app = next({ dev });
 const handle = app.getRequestHandler();
-const { createRouteMiddleware } = require('@paralleldrive/react-feature-toggles');
+const {
+  createExpressMiddleware
+} = require('@paralleldrive/react-feature-toggles');
 const features = require('./features');
 
-const createHandler = createRouteMiddleware(features);
+const createHandler = createExpressMiddleware(features);
 
 app.prepare().then(() => {
   const server = express();
 
   // This will check the feature and set the correct status code for this route.
   // 200 if the feature is enabled, otherwise it will be 404.
-  server.use('/profile', createHandler({
-    requiredFeature: 'profile',
-    get: (req, res) => {
-      return app.render(req, res, '/profile', req.query);
-    }
-  }));
+  server.use(
+    '/profile',
+    createHandler({
+      requiredFeature: 'profile',
+      get: (req, res) => {
+        return app.render(req, res, '/profile', req.query);
+      }
+    })
+  );
 
   server.get('*', (req, res) => {
     return handle(req, res);
@@ -31,4 +36,3 @@ app.prepare().then(() => {
     console.log(`> Ready on http://localhost:${port}`); // eslint-disable-line
   });
 });
-
